clean comment from all files and add extension responsible func to core data, done make three requests at same time after on boarding
