+++ NewsApp/Models/ArticlesModel.swift
@@ -40,18 +40,8 @@ struct Article: Codable {
 struct Source: Codable {
     let id: String?
     let name: String?
-//    let description: String
-//    let url: String
-//    let category: String
-//    let language: String
-//    let country: String
     enum CodingKeys: String, CodingKey {
         case id = "id"
         case name = "name"
-//        case description = "description"
-//        case url = "url"
-//        case category = "category"
-//        case language = "language"
-//        case country = "country"
     }
 }
