+++ NewsApp/Network/APIClint.swift
@@ -24,21 +24,16 @@ class APIClint:NetworkServiceProtocol {
             compeletion(.failure(.urlBadFormmated))
             return
         }
-        
         guard let urlRequest = URL(string: urlString) else {
             compeletion(.failure(.InternalError))
             return
         }
-        
         var  request = URLRequest(url: urlRequest)
         request.allHTTPHeaderFields = ["Content-Type": "application/json"]
         request.httpMethod = "\(method)"
         callNetwork(urlRequest: request, completion: compeletion)
-       
-        
     }
     func callNetwork<T:Codable>(urlRequest:URLRequest, completion: @escaping (Result<T, ErrorType>) -> Void) {
-        
         URLSession.shared.dataTask(with: urlRequest) { data, response, error in
             guard error == nil else {
                 completion(.failure(.ServerError))
@@ -48,16 +43,13 @@ class APIClint:NetworkServiceProtocol {
                 completion(.failure(.ServerError))
                 return
             }
-            
             do{
                 let object = try JSONDecoder().decode(T.self, from: data)
                 completion(.success(object))
-                print("response\(object)")
                 
             }    catch {
-                print((response as! HTTPURLResponse).statusCode)
+              //  print((response as! HTTPURLResponse).statusCode)
                     completion(.failure(.parsingError))
-
                 }
         }.resume()
     }
