+++ NewsApp/Network/NetworkServiceProtocol.swift
@@ -9,5 +9,4 @@ import Foundation
 protocol NetworkServiceProtocol{
     func getNews(countryName:String,catgoryId:String,completion:@escaping(Result<News,ErrorType>)->Void)
     func searchArticales(text:String,completion:@escaping(Result<News,ErrorType>)->Void)
-    
 }
