+++ NewsApp/Utilites/HandleConnection.swift
@@ -34,16 +34,3 @@ class HandelConnection{
     }
 
 }
-extension UIViewController{
-    func showAlertForInterNetConnection(){
-        let alert = UIAlertController(title: "network is not connected", message: "please, check your internet connection", preferredStyle: .alert)
-        let okBtn = UIAlertAction(title: "OK", style: .cancel, handler: nil)
-        alert.addAction(okBtn)
-        self.present(alert, animated: true, completion: nil)
-    }
-    func alertWarning(indexPath:IndexPath,title:String,message:String){
-        let alert = UIAlertController(title:title , message: message, preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "OK", style: .destructive))
-        self.present(alert, animated: true, completion: nil)
-    }
-}
