+++ NewsApp/Utilites/Utilities.swift
@@ -30,7 +30,6 @@ class Utilities{
         return UserDefaults.standard.value(forKey: "catgory") as? Array<Any> ?? []
     }
     func showMessage(message:String, error:Bool){
-        
         let view = MessageView.viewFromNib(layout: .messageView)
         if error == true {
             view.configureTheme(.error)
@@ -40,9 +39,16 @@ class Utilities{
         view.button?.isHidden = true
         view.titleLabel?.isHidden = true
         view.bodyLabel?.text = message
-        
         var config = SwiftMessages.Config()
         config.presentationStyle = .top
         SwiftMessages.show(config: config, view: view)
     }
+}
+extension UIViewController{
+    func alertWarning(indexPath:IndexPath,title:String,message:String){
+        let alert = UIAlertController(title:title , message: message, preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "OK", style: .destructive))
+        self.present(alert, animated: true, completion: nil)
+    }
+    
 }
