+++ NewsApp/Views/Home/ListNewsTb+Ex.swift
@@ -38,25 +38,12 @@ extension ListNewsViewController: UITableViewDelegate,UITableViewDataSource{
                 cell.source.text = arrayOfArticle[indexPath.row].source?.name
                 cell.date.text = arrayOfArticle[indexPath.row].publishedAt
                 cell.addToFav={ [self] in
-                    newsViewModel?.checkNewsInCoreData(data: arrayOfArticle[indexPath.row].publishedAt ?? "title", auther: arrayOfArticle[indexPath.row].author ?? "auther")
+                    checkIsFav(indexPath: indexPath)
                     isFav = newsViewModel?.isFav
                     if (isFav!) {
                         self.showDeleteAlert(indexPath: indexPath)
                     }else{
-                        do{
-                            try newsViewModel?.saveoCoreData(title: arrayOfArticle[indexPath.row].title ?? "title", img: arrayOfArticle[indexPath.row].urlToImage ?? "news", desc:arrayOfArticle[indexPath.row].description ?? "des", source: arrayOfArticle[indexPath.row].source?.name ?? "source", data: arrayOfArticle[indexPath.row].publishedAt ?? "2022/9/23", auther: arrayOfArticle[indexPath.row].author ?? "auther", completion: {  result in
-                                  switch result{
-                                  case true:
-                                      Utilities.utilities.showMessage(message: "added to Favourite", error: false)
-                                      print("add to core ")
-                                  case false :
-                                      print("faild to add to core")
-                                  }
-                              })
-                          }
-                      catch let error{
-                              print(error.localizedDescription)
-                          }
+                        saveToCoreData(indexPath: indexPath)
                     }
                 }
