+++ NewsApp/Views/Search/SearchViewController.swift
@@ -14,7 +14,7 @@ class SearchViewController: UIViewController,UISearchBarDelegate {
     
     @IBOutlet weak var searchTb: UITableView!
     var newsViewModel: NewsProtocolViewModel?
-    let disBag = DisposeBag()
+    fileprivate let disBag = DisposeBag()
     var arrayOfArticle: [Article] = []
     private  var isConn:Bool = false
     var articles : [Articles] = []
@@ -32,8 +32,6 @@ class SearchViewController: UIViewController,UISearchBarDelegate {
         searchBar.delegate=self
         refreshController.addTarget(self, action: #selector(pullToRefresh), for: .valueChanged)
         searchTb.addSubview(refreshController)
-      //  self.searchTb.isHidden=false
-
     }
     
     func searchRequest(){
@@ -43,7 +41,6 @@ class SearchViewController: UIViewController,UISearchBarDelegate {
             .subscribe { news in
                 if news.count == 0 {
                     self.noSershView.isHidden=false
-                    
                 }
                 else{
                     self.noSershView.isHidden=true
@@ -102,50 +99,4 @@ class SearchViewController: UIViewController,UISearchBarDelegate {
     */
 
 }
-extension SearchViewController{
-    func getArticlesFromCoreData(){
-        do{
-            try  newsViewModel?.getAllnewsInCoreData(completion: { response in
-                //MARK: LSA M5LST4
-                switch response{
-                case true:
-                    print("data retrived successfuly")
-                case false:
-                    print("data cant't retrieved")
-                }
-            })
-        }
-        catch let error{
-            print(error.localizedDescription)
-        }
-        articles = (newsViewModel?.articleList)!
-        searchTb.reloadData()
-            }
-    func deleteItemFromCoreData(index:IndexPath){
-        do{
-            try self.newsViewModel?.removeNewsFromCoreDatat(title: "\(arrayOfArticle[index.row].title ?? "title")", completionHandler: { result in
-                switch result{
-                case true:
-                print("remove from cart")
-                self.getArticlesFromCoreData()
-                Utilities.utilities.showMessage(message: "remove from Favourite", error: false)
-                self.searchTb.reloadData()
-                case false:
-                    print("cann't delet")
-                        }
-            })
-        }
-        catch let error{
-            print(error.localizedDescription)
-        }
-    }
-    func
-    showDeleteAlert(indexPath:IndexPath){
-        let alert = UIAlertController(title: "Already Saved", message: "Are You remove this item from the Fav", preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "Cancel", style: .default, handler: nil))
-        alert.addAction(UIAlertAction(title: "OK", style: .destructive, handler: { [self] UIAlertAction in
-        self.deleteItemFromCoreData(index: indexPath)
-        }))
-        self.present(alert, animated: true, completion: nil)
-    }
-}
+
